requirements.txt
Django==1.11.2
Pillow==4.1.1
django-storages
djangorestframework==3.6.3
django-cors-middleware
django-rest-auth
django-extra-fields
django-crispy-forms==1.6.1
django-extensions==1.7.9
django-multiselectfield
django-filter
django-dotenv
gunicorn==19.7.1
psycopg2==2.7.1
redis==2.10.5
sorl-thumbnail==12.4a1
sorl-thumbnail-serializer-field==0.1
beautifulsoup4
# whitenoise==3.3.0
brotlipy
requests
ipdb
coreapi
boto
